Bugfix, move function
