+++ ansible/roles/wordpress-instance/action_plugins/wordpress_action_module.py
@@ -6,6 +6,7 @@ from ansible.plugins.action import ActionBase
 from ansible.errors import AnsibleActionFail
 from ansible.module_utils import six
 
+import re
 
 class WordPressActionModule(ActionBase):
 
@@ -71,6 +72,18 @@ class WordPressActionModule(ActionBase):
             return None
         else:
             return self._templar.template(unexpanded)
+
+    def _is_filename (self, from_piece):
+        """
+        Tells if a path is a filename or not.
+
+        :param from_piece: string describing plugin source.
+        """
+        self._log(re.match(r'^https:\/\/github\.com\/[\w-]+\/[\w-]+(\/)?$', from_piece))
+        return (from_piece != "wordpress.org/plugins"
+                # check if match a github repo
+                and not re.match(r'^https:\/\/github\.com\/[\w-]+\/[\w-]+(\/)?$', from_piece)
+                and not from_piece.endswith(".zip"))
 
 
     def _update_result (self, result):
